requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.10.6
pandas==2.2.3
python-dateutil==2.8.2